Make Alembic the single, working schema mechanism
The schema was built by Base.metadata.create_all() at startup while a full
Alembic setup existed but was never run (no alembic_version in the live DB).
create_all only creates missing tables — it never ALTERs/drops — so the
schema couldn't evolve reproducibly and two sources of truth disagreed.

- Squash the 6-migration chain into one clean 0001 baseline reflecting the
  current models only (3 flow tables, no api_keys — that table is never
  created now).
- Fix Alembic wiring: script_location = %(here)s/alembic (cwd-independent),
  and env.py reuses connection._resolve_database_url() so Alembic and the app
  target the same DB (same postgres:// rewrite + absolute SQLite path).
- Remove create_all from main.py; Alembic is now the single source of truth.
- Run `alembic upgrade head` in the render.yaml startCommand (free plan has no
  preDeployCommand; free = single instance, so no migration race). `&&` fails
  the boot loudly on a bad migration.
- README: document the one-time local `alembic upgrade head` step and that
  Render runs migrations automatically on deploy.

Verified locally on fresh SQLite: upgrade -> correct tables + alembic_version
0001, no api_keys; downgrade base -> upgrade round-trip; the flows-endpoint DB
path works against the migrated schema with create_all removed.
