+++ app/backend/alembic.ini
@@ -2,8 +2,11 @@
 
 [alembic]
 # path to migration scripts
-# Use forward slashes (/) also on windows to provide an os agnostic path
-script_location = alembic
+# Use forward slashes (/) also on windows to provide an os agnostic path.
+# %(here)s resolves to this ini file's own directory, so `alembic -c
+# app/backend/alembic.ini ...` works regardless of the current working
+# directory (e.g. repo root locally and on Render).
+script_location = %(here)s/alembic
 
 # template used to generate migration file names; The default value is %%(rev)s_%%(slug)s
 # Uncomment the line below if you want the files to be prepended with date and time
