+++ app/backend/alembic/versions/0001_initial_schema.py
@@ -1,0 +1,99 @@
+"""initial schema
+
+Revision ID: 0001
+Revises: 
+Create Date: 2026-07-07 09:03:50.986104
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '0001'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('hedge_fund_flows',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('(CURRENT_TIMESTAMP)'), nullable=True),
+    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=True),
+    sa.Column('name', sa.String(length=200), nullable=False),
+    sa.Column('description', sa.Text(), nullable=True),
+    sa.Column('nodes', sa.JSON(), nullable=False),
+    sa.Column('edges', sa.JSON(), nullable=False),
+    sa.Column('viewport', sa.JSON(), nullable=True),
+    sa.Column('data', sa.JSON(), nullable=True),
+    sa.Column('is_template', sa.Boolean(), nullable=True),
+    sa.Column('tags', sa.JSON(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_hedge_fund_flows_id'), 'hedge_fund_flows', ['id'], unique=False)
+    op.create_table('hedge_fund_flow_runs',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('flow_id', sa.Integer(), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('(CURRENT_TIMESTAMP)'), nullable=True),
+    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=True),
+    sa.Column('status', sa.String(length=50), nullable=False),
+    sa.Column('started_at', sa.DateTime(timezone=True), nullable=True),
+    sa.Column('completed_at', sa.DateTime(timezone=True), nullable=True),
+    sa.Column('trading_mode', sa.String(length=50), nullable=False),
+    sa.Column('schedule', sa.String(length=50), nullable=True),
+    sa.Column('duration', sa.String(length=50), nullable=True),
+    sa.Column('request_data', sa.JSON(), nullable=True),
+    sa.Column('initial_portfolio', sa.JSON(), nullable=True),
+    sa.Column('final_portfolio', sa.JSON(), nullable=True),
+    sa.Column('results', sa.JSON(), nullable=True),
+    sa.Column('error_message', sa.Text(), nullable=True),
+    sa.Column('run_number', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['flow_id'], ['hedge_fund_flows.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_hedge_fund_flow_runs_flow_id'), 'hedge_fund_flow_runs', ['flow_id'], unique=False)
+    op.create_index(op.f('ix_hedge_fund_flow_runs_id'), 'hedge_fund_flow_runs', ['id'], unique=False)
+    op.create_table('hedge_fund_flow_run_cycles',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('flow_run_id', sa.Integer(), nullable=False),
+    sa.Column('cycle_number', sa.Integer(), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('(CURRENT_TIMESTAMP)'), nullable=True),
+    sa.Column('started_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('completed_at', sa.DateTime(timezone=True), nullable=True),
+    sa.Column('analyst_signals', sa.JSON(), nullable=True),
+    sa.Column('trading_decisions', sa.JSON(), nullable=True),
+    sa.Column('executed_trades', sa.JSON(), nullable=True),
+    sa.Column('portfolio_snapshot', sa.JSON(), nullable=True),
+    sa.Column('performance_metrics', sa.JSON(), nullable=True),
+    sa.Column('status', sa.String(length=50), nullable=False),
+    sa.Column('error_message', sa.Text(), nullable=True),
+    sa.Column('llm_calls_count', sa.Integer(), nullable=True),
+    sa.Column('api_calls_count', sa.Integer(), nullable=True),
+    sa.Column('estimated_cost', sa.String(length=20), nullable=True),
+    sa.Column('trigger_reason', sa.String(length=100), nullable=True),
+    sa.Column('market_conditions', sa.JSON(), nullable=True),
+    sa.ForeignKeyConstraint(['flow_run_id'], ['hedge_fund_flow_runs.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_hedge_fund_flow_run_cycles_flow_run_id'), 'hedge_fund_flow_run_cycles', ['flow_run_id'], unique=False)
+    op.create_index(op.f('ix_hedge_fund_flow_run_cycles_id'), 'hedge_fund_flow_run_cycles', ['id'], unique=False)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f('ix_hedge_fund_flow_run_cycles_id'), table_name='hedge_fund_flow_run_cycles')
+    op.drop_index(op.f('ix_hedge_fund_flow_run_cycles_flow_run_id'), table_name='hedge_fund_flow_run_cycles')
+    op.drop_table('hedge_fund_flow_run_cycles')
+    op.drop_index(op.f('ix_hedge_fund_flow_runs_id'), table_name='hedge_fund_flow_runs')
+    op.drop_index(op.f('ix_hedge_fund_flow_runs_flow_id'), table_name='hedge_fund_flow_runs')
+    op.drop_table('hedge_fund_flow_runs')
+    op.drop_index(op.f('ix_hedge_fund_flows_id'), table_name='hedge_fund_flows')
+    op.drop_table('hedge_fund_flows')
+    # ### end Alembic commands ###
