--- app/backend/alembic/versions/add_api_keys_table.py
@@ -1,44 +1,0 @@
-"""add_api_keys_table
-
-Revision ID: d5e78f9a1b2c
-Revises: 3f9a6b7c8d2e
-Create Date: 2025-07-27 15:20:00.000000
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'd5e78f9a1b2c'
-down_revision: Union[str, None] = '3f9a6b7c8d2e'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # Create API keys table
-    op.create_table('api_keys',
-        sa.Column('id', sa.Integer(), nullable=False),
-        sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('(CURRENT_TIMESTAMP)'), nullable=True),
-        sa.Column('updated_at', sa.DateTime(timezone=True), nullable=True),
-        sa.Column('provider', sa.String(length=100), nullable=False),
-        sa.Column('key_value', sa.Text(), nullable=False),
-        sa.Column('is_active', sa.Boolean(), nullable=True),
-        sa.Column('description', sa.Text(), nullable=True),
-        sa.Column('last_used', sa.DateTime(timezone=True), nullable=True),
-        sa.PrimaryKeyConstraint('id'),
-        sa.UniqueConstraint('provider')
-    )
-    op.create_index(op.f('ix_api_keys_id'), 'api_keys', ['id'], unique=False)
-    op.create_index(op.f('ix_api_keys_provider'), 'api_keys', ['provider'], unique=False)
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    op.drop_index(op.f('ix_api_keys_provider'), table_name='api_keys')
-    op.drop_index(op.f('ix_api_keys_id'), table_name='api_keys')
-    op.drop_table('api_keys')
